finished base structure

+++ Praktikum05/src/main/java/org/swa/assignments/al/ModifyAssignment.java
@@ -5,6 +5,6 @@ import org.swa.boundary.entity.AssignmentDTO;
 public interface ModifyAssignment {
 
     void editAssignment(AssignmentDTO assignmentDTO);
-    boolean deleteAssignment(int id);
+    boolean deleteAssignment(long id);
 
 }
